Fixes for Amazon Orders
