+++ src/extractor/amazonOrders/models/amazonBaseModel.ts
