+++ src/extractor/index.ts
@@ -1,3 +1,4 @@
+import * as AmazonOrders from "./amazonOrders/index.js";
 import * as Facebook from "./facebook/index.js";
 import * as Instagram from "./instagram/index.js";
 import * as NetflixViewingHistory from "./netflixViewingHistory/index.js";
@@ -9,4 +10,5 @@ Facebook.initialize();
 Instagram.initialize();
 NetflixViewingHistory.initialize();
 Youtube.initialize();
+AmazonOrders.initialize();
 export const register = () => {};
