+++ test/data/csv/amazon_orders/01-Aug-2022_to_07-Oct-2022.csv
@@ -1,0 +1,14 @@
+Order Date,Order ID,Title,Category,ASIN/ISBN,UNSPSC Code,Website,Release Date,Condition,Seller,Seller Credentials,List Price Per Unit,Purchase Price Per Unit,Quantity,Payment Instrument Type,Purchase Order Number,PO Line Number,Ordering Customer Email,Shipment Date,Shipping Address Name,Shipping Address Street 1,Shipping Address Street 2,Shipping Address City,Shipping Address State,Shipping Address Zip,Order Status,Carrier Name & Tracking Number,Item Subtotal,Item Subtotal Tax,Item Total,Tax Exemption Applied,Tax Exemption Type,Exemption Opt-Out,Buyer Name,Currency,Group Name
+08/01/22,113-8557626-0722622,"iHealth COVID-19 Antigen Rapid Test, 2 Tests per Pack,FDA EUA Authorized OTC At-home Self Test, Results in 15 Minutes with Non-invasive Nasal Swab, Easy to Use & No Discomfort",CLINICAL_DIAGNOSTIC_TESTING_KIT,"B09KZ6TBNY","42000000",Amazon.com,,new,iHealthLabs,889 certification,$17.98,$17.98,4,"MasterCard - 7216",,,services@vana.org,08/01/22,Jens Midthun,416 S SPRING ST APT 1209,,LOS ANGELES,CA,90013-1959,Shipped,AMZN_US(TBA034658933103),$71.92,$6.84,$78.76,false,,false,Vana Services,USD,
+08/02/22,114-4696448-1009031,"Wet Ones Antibacterial Hand Wipes, Ocean Breeze, 50 Count",BEAUTY,"B093NWF5RR","53131600",Amazon.com,,new,Amazon.com,,$4.99,$4.99,1,"MasterCard - 5359",,,services@vana.org,08/02/22,Joseph Gillespie,138 CALLE COLORADO,,SAN DIMAS,CA,91773-3994,Shipped,AMZN_US(TBA197700187904),$4.99,$0.47,$5.46,false,,false,Vana Services,USD,
+08/02/22,114-4696448-1009031,"Nature’s Rite Sinus Relief™, 1.5 oz. (45 mL) – All-Natural Nasal Sinus Spray – Relief for Sinus Problems – Encourages a Healthy Sinus System – Easy-to-Use, Portable Sinus Support – Made in USA",MEDICATION,"B00E6D1E8Y","51000000",Amazon.com,,new,Active Recovery Essentials,,$0.00,$10.99,1,"MasterCard - 5359",,,services@vana.org,08/02/22,Joseph Gillespie,138 CALLE COLORADO,,SAN DIMAS,CA,91773-3994,Shipped,AMZN_US(TBA197700187904),$10.99,$1.04,$12.03,false,,false,Vana Services,USD,
+08/02/22,114-4696448-1009031,"Alcedo Pulse Oximeter Fingertip | Blood Oxygen Saturation Level (SpO2) and Heart Rate Monitor | Dual Color OLED Display | Portable Carry Case, Lanyard, Batteries",BLOOD_OXYGEN_MONITOR,"B086CBYN87","42181801",Amazon.com,,new,Alcedo Health,,$26.95,$15.96,1,"MasterCard - 5359",,,services@vana.org,08/02/22,Joseph Gillespie,138 CALLE COLORADO,,SAN DIMAS,CA,91773-3994,Shipped,AMZN_US(TBA197700187904),$15.96,$1.52,$17.48,false,,false,Vana Services,USD,
+08/02/22,114-4696448-1009031,"iHealth Digital Oral Thermometer PT1,Fever Thermometer with Dual-Sensors for High Accuracy, Rectum Armpit Reading Thermometer for Adults and Babies, Memory Recall and Extra Large LCD Backlit Display",THERMOMETER,"B08L4BNY5Z","42182200",Amazon.com,,new,iHealthLabs,889 certification,$14.99,$10.99,1,"MasterCard - 5359",,,services@vana.org,08/02/22,Joseph Gillespie,138 CALLE COLORADO,,SAN DIMAS,CA,91773-3994,Shipped,AMZN_US(TBA197700187904),$10.99,$1.04,$12.03,false,,false,Vana Services,USD,
+08/02/22,114-4696448-1009031,"iHealth COVID-19 Antigen Rapid Test, 2 Tests per Pack,FDA EUA Authorized OTC At-home Self Test, Results in 15 Minutes with Non-invasive Nasal Swab, Easy to Use & No Discomfort",CLINICAL_DIAGNOSTIC_TESTING_KIT,"B09KZ6TBNY","42000000",Amazon.com,,new,iHealthLabs,889 certification,$17.98,$17.98,2,"MasterCard - 5359",,,services@vana.org,08/02/22,Joseph Gillespie,138 CALLE COLORADO,,SAN DIMAS,CA,91773-3994,Shipped,AMZN_US(TBA197376454704),$35.96,$3.42,$39.38,false,,false,Vana Services,USD,
+08/02/22,114-4696448-1009031,"Traditional Medicinals Organic Echinacea Plus Elderberry Herbal Tea, Promotes Immune Function, (Pack of 2) - 32 Tea Bags Total",TEA,"B001F1XCDK","50201713",Amazon.com,,new,TheNewMall,,$16.24,$16.38,1,"MasterCard - 5359",,,services@vana.org,08/02/22,Joseph Gillespie,138 CALLE COLORADO,,SAN DIMAS,CA,91773-3994,Shipped,AMZN_US(TBA197700187904),$16.38,$0.00,$16.38,,,,Vana Services,USD,
+08/02/22,114-4696448-1009031,"BEEKEEPER'S NATURALS Propolis Throat Spray - 95% Bee Propolis Extract - Natural Immune Support & Sore Throat Relief - Antioxidants, Keto, Paleo, Gluten-Free (1.06 oz)",MEDICATION,"B019ND0DU6","53131600",Amazon.com,,new,Beekeeper's Naturals,,$13.99,$13.95,1,"MasterCard - 5359",,,services@vana.org,08/02/22,Joseph Gillespie,138 CALLE COLORADO,,SAN DIMAS,CA,91773-3994,Shipped,AMZN_US(TBA197700187904),$13.95,$1.33,$15.28,false,,false,Vana Services,USD,
+08/02/22,114-4696448-1009031,"EasyEast N95 Mask Respirator [ Made in USA ] NIOSH Certified N95 Particulate Respirators Face Mask (Pack of 10),White,Adult,501831 (EE10P)",RESPIRATOR,"B08ZHRTS52","46182008",Amazon.com,,new,Amazon.com,,$29.99,$21.01,1,"MasterCard - 5359",,,services@vana.org,08/02/22,Joseph Gillespie,138 CALLE COLORADO,,SAN DIMAS,CA,91773-3994,Shipped,AMZN_US(TBA197700187904),$21.01,$2.00,$23.01,false,,false,Vana Services,USD,
+08/24/22,113-0865964-5049827,Apple USB-C to Magsafe 3 Cable (2 m),ELECTRONIC_CABLE,"B09JR6YFQT","43202222",Amazon.com,,new,Amazon.com,,$49.00,$49.00,1,"MasterCard - 5359",,,services@vana.org,08/25/22,Jenny Quan,28 S HILL BLVD,,SAN FRANCISCO,CA,94112-4549,Shipped,AMZN_US(TBA027505191104),$49.00,$4.23,$53.23,false,,false,Vana Services,USD,
+08/24/22,113-0865964-5049827,"140W USB C Charger for MacBook Pro 16 inch 2021, MacBook Air, USB-C Power Adapter for iPhone 13/12, iPad Pro/Air, Galaxy S22, Google Pixelbook, Pixel, ThinkPad, Dell XPS",CHARGING_ADAPTER,"B0B3LKB73Y","26111704",Amazon.com,,new,IFEART,,$79.99,$79.99,1,"MasterCard - 5359",,,services@vana.org,08/25/22,Jenny Quan,28 S HILL BLVD,,SAN FRANCISCO,CA,94112-4549,Shipped,AMZN_US(TBA028235167204),$79.99,$6.04,$86.03,false,,false,Vana Services,USD,
+08/25/22,112-1760062-5747452,Play Bigger,ABIS_BOOK,"0349411360","55101500",Amazon.com,2016-06-14T07:00:01,new,Amazon.com,,$19.99,$22.02,1,"MasterCard - 5359",,,services@vana.org,08/25/22,Lauren Barrett,213 EASTERN PKWY APT 6,,BROOKLYN,NY,11238-6109,Shipped,AMZN_US(TBA028722115804),$22.02,$1.95,$23.97,false,,false,Vana Services,USD,
+09/01/22,113-1554413-4993034,"1080P Webcam with Microphone, EMEET C960 Web Camera, 2 Mics Streaming Webcam with Privacy Cover, 90°View Computer Camera, Plug&Play USB Webcam for Calls/Conference, Zoom/Skype/YouTube, Laptop/Desktop",CAMCORDER,"B07M6Y7355","45121520",Amazon.com,,new,eMeetDirect,ISO 9001,$39.99,$34.99,1,"MasterCard - 7216",,,services@vana.org,09/01/22,Ashley Haylett,599 JACKSON ST APT B,,ALBANY,CA,94706-1150,Shipped,AMZN_US(TBA037986461504),$34.99,$3.39,$38.38,false,,false,Vana Services,USD,
