--- test/data/csv/netflix_light/NetflixViewingHistory.csv
@@ -1,3 +1,0 @@
-Title,Date
-Our Planet: Season 1: Forests,2/15/20
-Our Planet - Behind The Scenes,2/15/20
