+++ test/utils/fileToZip.test.ts
@@ -18,27 +18,46 @@ describe("utils/fileToZip", () => {
   });
 
   it("fileToZip() returns a ServiceFile", async () => {
-    deleteFile(
-      `csv/${config.SERVICE_NETFLIX_VIEWING_HISTORY}/${config.SERVICE_NETFLIX_VIEWING_HISTORY}.sqlite`
-    );
-    const file = await loadTestFile(
-      `csv/${config.SERVICE_NETFLIX_VIEWING_HISTORY}/NetflixViewingHistory.csv`
-    );
-    const serviceFile = await filesToZip(
-      config.SERVICE_NETFLIX_VIEWING_HISTORY,
-      [file]
-    );
-    const database = await zipToSQLiteInstance([serviceFile], false);
+    const testFiles = {
+      [config.SERVICE_NETFLIX_VIEWING_HISTORY]: ["NetflixViewingHistory"],
+      [config.SERVICE_AMAZON_ORDERS]: [
+        "01-Aug-2022_to_07-Oct-2022",
+        "01-Jan-2010_to_08-Oct-2010",
+        "Amazon_Orders_csv",
+      ],
+    };
 
-    saveSqliteDump(
-      database.exportDatabase(config.SERVICE_NETFLIX_VIEWING_HISTORY),
-      `csv/${config.SERVICE_NETFLIX_VIEWING_HISTORY}/${config.SERVICE_NETFLIX_VIEWING_HISTORY}.sqlite`
-    );
-    database.close();
-    expect(
-      fileExists(
-        `csv/${config.SERVICE_NETFLIX_VIEWING_HISTORY}/${config.SERVICE_NETFLIX_VIEWING_HISTORY}.sqlite`
-      )
-    ).toBeTruthy();
+    /**
+     * Loads a CSV file, structures it and ensures the generated .sqlite file exists
+     * @param serviceName (ie: instagram)
+     * @param testFile (ie: instagram_json_1)
+     */
+    const runTestAgainstFile = async (
+      serviceName: string,
+      testFile: string
+    ): Promise<void> => {
+      deleteFile(`csv/${serviceName}/${testFile}.sqlite`);
+      const file = await loadTestFile(`csv/${serviceName}/${testFile}.csv`);
+      const serviceFile = await filesToZip(serviceName, [file]);
+      const database = await zipToSQLiteInstance([serviceFile], false);
+
+      saveSqliteDump(
+        database.exportDatabase(serviceName),
+        `csv/${serviceName}/${testFile}.sqlite`
+      );
+      database.close();
+      expect(fileExists(`csv/${serviceName}/${testFile}.sqlite`)).toBeTruthy();
+    };
+
+    const testPromises: Promise<void>[] = [];
+    Object.entries(testFiles).forEach((service: [string, string[]]) => {
+      const serviceName = service[0];
+      const testFilesForService = service[1];
+      testFilesForService.forEach((testFile: string) => {
+        testPromises.push(runTestAgainstFile(serviceName, testFile));
+      });
+    });
+
+    await Promise.all(testPromises);
   });
 });
